RoleRadar/requirements.txt
requests
pandas
matplotlib
seaborn
joblib
scikit-learn
Django
protobuf
RapidFuzz
numpy
pathlib
#Markdown
xhtml2pdf
python-docx
beautifulsoup4
#markdownify
django-markdownify
#genai
#google-generativeai
google-genai
fitz
PyMuPDF